YAHOU

+++ ft_striteri.c
@@ -6,7 +6,7 @@
 /*   By: aviala <aviala@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2013/11/23 20:16:48 by aviala            #+#    #+#             */
-/*   Updated: 2013/11/25 17:45:48 by aviala           ###   ########.fr       */
+/*   Updated: 2013/11/26 10:12:33 by aviala           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,7 +15,7 @@
 
 void ft_striteri(char *s, void (*f)(unsigned int, char *))
 {
-	unsigned int	i;
+	size_t	i;
 
 	i = 0;
 	while (s[i])
